Add maintenance components and refactor page
Introduce reusable maintenance UI: adds maintenanceHelpers (constants, styles, formatters), Modal, MaintenanceBadges (Toast, Loader, Empty, badges, Pager, Field, Row2), CreateMaintenanceModal and DetailMaintenanceModal components. Refactors Agent Maintenance page to import and use these new components and helpers (removes in-file duplicates and centralizes styling/logic), and wires up create/detail modals, pager and toast handling. This modularizes the maintenance feature and reduces duplication.
